refactor: Remove CSS modules from components
In order to simplify the styling and reduce duplication, the individual
CSS modules for the components have been removed. Since the styles
are consistent across the inputs, a new and unique CSS module has been
created. This allows for centralized styling while still providing the
option to reserve specific styles if necessary.